refactor: drop ontology/runtime barrels and unused engine setup
Point callers at src/core directly and remove the unused src/engine/setup helpers that oms setup never invoked.
