+++ docs/exec-plan/archived/self-owned-second-brain/deep-interview-record.md
