+++ docs/exec-plan/archived/self-owned-second-brain/plan.md
