+++ docs/exec-plan/archived/self-owned-second-brain/spec.md
