+++ src/graph/cache.ts
@@ -4,8 +4,8 @@ import path from "node:path";
 import { parseNote } from "../conventions/frontmatter.js";
 import { validateFrontmatter } from "../conventions/validate.js";
 import { safeVaultNotePath } from "../capture/safe.js";
-import { resolveConcept } from "../ontology/resolver.js";
-import type { Concept, Ontology } from "../ontology/types.js";
+import { resolveConcept } from "../core/ontology/resolver.js";
+import type { Concept, Ontology } from "../core/ontology/types.js";
 
 export type GraphEdgeType = "folder-concept" | "property-axis" | "property-value" | "wikilink";
