+++ src/ontology/loader.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import { loadOntology } from "./loader.js";
+import { loadOntology } from "../core/ontology/loader.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
