--- src/ontology/loader.ts
@@ -1,1 +1,0 @@
-export * from "../core/ontology/loader.js";
