--- src/ontology/resolver.ts
@@ -1,1 +1,0 @@
-export * from "../core/ontology/resolver.js";
