--- src/ontology/types.ts
@@ -1,1 +1,0 @@
-export * from "../core/ontology/types.js";
