--- src/runtime/assets.ts
@@ -1,1 +1,0 @@
-export * from "../core/runtime/assets.js";
